MovingTarget
Made a moving target that goes from initial position straight to the *endingPosition* at a certain *speed*, and reverse.
Implemented Target health. Ammunition will decrease health and object will be destroied when health reaches cero.
IsHitedBy and OnCollisionEnter upgraded to Target class instead of FixedTarget only.

+++ Assets/Scripts/Ammunition.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 [RequireComponent(typeof(Rigidbody))]
 public abstract class Ammunition : MonoBehaviour
 {
-    public float dammage = 1;
+    public int dammage = 1;
     protected Rigidbody rb;
 
     private void Awake()
